fix remaining 3d panel cycle issues
