+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/Markers.ts
@@ -7,7 +7,7 @@ import { set } from "lodash";
 import { toNanoSec } from "@foxglove/rostime";
 import { SettingsTreeAction } from "@foxglove/studio";
 
-import { LayerSettingsMarkerNamespace, TopicMarkers } from "./TopicMarkers";
+import { LayerSettingsMarker, LayerSettingsMarkerNamespace, TopicMarkers } from "./TopicMarkers";
 import type { IRenderer } from "../IRenderer";
 import { SELECTED_ID_VARIABLE } from "../Renderable";
 import { PartialMessage, PartialMessageEvent, SceneExtension } from "../SceneExtension";
@@ -22,16 +22,8 @@ import {
   normalizeVector3s,
 } from "../normalizeMessages";
 import { Marker, MarkerArray, MARKER_ARRAY_DATATYPES, MARKER_DATATYPES } from "../ros";
-import { BaseSettings } from "../settings";
 import { topicIsConvertibleToSchema } from "../topicIsConvertibleToSchema";
 import { makePose } from "../transforms";
-
-export type LayerSettingsMarker = BaseSettings & {
-  color: string | undefined;
-  showOutlines: boolean | undefined;
-  selectedIdVariable: string | undefined;
-  namespaces: Record<string, LayerSettingsMarkerNamespace>;
-};
 
 const DEFAULT_SETTINGS: LayerSettingsMarker = {
   visible: false,
