+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/markers/RenderableMarker.ts
@@ -11,7 +11,7 @@ import type { IRenderer } from "../../IRenderer";
 import { BaseUserData, Renderable } from "../../Renderable";
 import { makeRgba, rgbToThreeColor, stringToRgba } from "../../color";
 import { Marker } from "../../ros";
-import { LayerSettingsMarker } from "../Markers";
+import type { LayerSettingsMarker } from "../TopicMarkers";
 
 const tempColor = new THREE.Color();
 const tempColor2 = new THREE.Color();
